+++ src/main/kotlin/io/vlang/ide/run/VlangRunConfiguration.kt
@@ -1,159 +1,97 @@
 package io.vlang.ide.run
 
 import com.intellij.execution.Executor
-import com.intellij.execution.configurations.ConfigurationFactory
-import com.intellij.execution.configurations.LocatableConfiguration
-import com.intellij.execution.configurations.RunConfigurationBase
-import com.intellij.execution.configurations.RunProfileState
+import com.intellij.execution.configurations.*
 import com.intellij.execution.runners.ExecutionEnvironment
+import com.intellij.openapi.actionSystem.DataContext
+import com.intellij.openapi.options.ConfigurationQuickFix
 import com.intellij.openapi.project.Project
-import org.jdom.Element
+import com.intellij.openapi.project.guessProjectDir
+import com.intellij.openapi.util.NlsActions
+import java.io.File
 
 open class VlangRunConfiguration(project: Project, factory: ConfigurationFactory?, name: String?) :
     RunConfigurationBase<VlangRunConfigurationOptions>(project, factory, name),
+    RunProfileWithCompileBeforeLaunchOption,
     LocatableConfiguration {
 
-    override fun getOptions() = super.getOptions() as VlangRunConfigurationOptions
-
-    var runKind: VlangRunConfigurationEditor.RunKind
-        get() = options.runKind
-        set(value) {
-            options.runKind = value
-        }
-
-    var fileName: String
-        get() = options.fileName
-        set(value) {
-            options.fileName = value
-        }
-
-    var directory: String
-        get() = options.directory
-        set(value) {
-            options.directory = value
-        }
-
-    var outputDir: String
-        get() = options.outputDir
-        set(value) {
-            options.outputDir = value
-        }
-
-    var runAfterBuild: Boolean
-        get() = options.runAfterBuild
-        set(value) {
-            options.runAfterBuild = value
-        }
-
-    var workingDir: String
-        get() = options.workingDir
-        set(value) {
-            options.workingDir = value
-        }
-
-    var envs: MutableMap<String, String>
-        get() = options.envs
-            .ifEmpty { null }
-            ?.split(";")
-            ?.map { it.split("=") }
-            ?.associate { it[0] to it[1] }
-            ?.toMutableMap() ?: mutableMapOf()
-        set(value) {
-            options.envs = value
-                .map { "${it.key}=${it.value}" }
-                .joinToString(";")
-        }
-
-    var buildArguments: String
-        get() = options.buildArguments
-        set(value) {
-            options.buildArguments = value
-        }
-
-    var programArguments: String
-        get() = options.programArguments
-        set(value) {
-            options.programArguments = value
-        }
-
-    var production: Boolean
-        get() = options.production
-        set(value) {
-            options.production = value
-        }
-
-    var emulateTerminal: Boolean
-        get() = options.emulateTerminal
-        set(value) {
-            options.emulateTerminal = value
-        }
-
-    override fun writeExternal(element: Element) {
-        super<RunConfigurationBase>.writeExternal(element)
-
-        with(element) {
-            writeString("runKind", runKind.name)
-            writeString("fileName", fileName)
-            writeString("directory", directory)
-            writeString("outputDir", outputDir)
-            writeBool("runAfterBuild", runAfterBuild)
-            writeString("workingDir", workingDir)
-            writeString("envs", options.envs)
-            writeString("buildArguments", buildArguments)
-            writeString("programArguments", programArguments)
-            writeBool("production", production)
-            writeBool("emulateTerminal", emulateTerminal)
-        }
-    }
+    public override fun getOptions() = super.getOptions() as VlangRunConfigurationOptions
 
     override fun isGeneratedName(): Boolean {
         return false
     }
 
-    override fun readExternal(element: Element) {
-        super<RunConfigurationBase>.readExternal(element)
-
-        with(element) {
-            readString("runKind")?.let { runKind = VlangRunConfigurationEditor.RunKind.fromString(it) }
-            readString("fileName")?.let { fileName = it }
-            readString("directory")?.let { directory = it }
-            readString("outputDir")?.let { outputDir = it }
-            readBool("runAfterBuild")?.let { runAfterBuild = it }
-            readString("workingDir")?.let { workingDir = it }
-            readString("envs")?.let { options.envs = it }
-            readString("buildArguments")?.let { buildArguments = it }
-            readString("programArguments")?.let { programArguments = it }
-            readBool("production")?.let { production = it }
-            readBool("emulateTerminal")?.let { emulateTerminal = it }
+    override fun suggestedName(): @NlsActions.ActionText String? {
+        val name = if (options.runKind == VlangRunConfigurationEditor.RunKind.File) {
+            File(options.fileName).nameWithoutExtension
+        } else {
+            val dir = File(options.directory).name
+            if (dir == ".") {
+                File(options.workingDir).name
+            } else {
+                dir
+            }
         }
+        return "V Build $name"
     }
 
     override fun getConfigurationEditor() = VlangRunConfigurationEditor(project)
 
+    @Suppress("UnstableApiUsage")
     override fun checkConfiguration() {
-        // TODO: add check for class or method?
+        // RuntimeConfigurationWarning – if the configuration settings contain a problem which the user should be warned about.
+        // RuntimeConfigurationException – if the configuration settings contain a non-fatal error which the user should be warned about but the execution should still be allowed.
+        // RuntimeConfigurationError – if the configuration settings contain a fatal error which makes it impossible to execute the run configuration.
+
+        val opt = this.expandedOptions()
+
+        if (opt.workingDir.isBlank()) {
+            throw RuntimeConfigurationError("Working directory is empty", object : ConfigurationQuickFix {
+                override fun applyFix(dataContext: DataContext) {
+                    options.workingDir = project.guessProjectDir()?.path ?: ""
+                }
+            })
+        }
+
+        if (opt.runKind == VlangRunConfigurationEditor.RunKind.File) {
+            if (opt.fileName.isBlank()) {
+                throw RuntimeConfigurationError("Source file path is empty")
+            }
+            if (!opt.fileName.endsWith(".v") && !opt.fileName.endsWith(".vv") && !opt.fileName.endsWith(".vsh") ) {
+                throw RuntimeConfigurationError("Source file must be a V file (\".v\", \".vv\", \".vsh\")")
+            }
+            if (!File(opt.workingDir, opt.fileName).exists()) {
+                throw RuntimeConfigurationError("Source file '${opt.fileName}' does not exist")
+            }
+        } else {
+            if (opt.directory.isBlank()) {
+                throw RuntimeConfigurationError("Source directory path is empty")
+            }
+            if (!File(opt.workingDir, opt.directory).exists()) {
+                throw RuntimeConfigurationError("Source directory '${opt.directory}' does not exist")
+            }
+        }
+
+        if (opt.buildArguments.contains("-o") || opt.buildArguments.contains("-output")) {
+            throw RuntimeConfigurationError("Output file is set in build arguments - please use 'Output file' field instead!", object : ConfigurationQuickFix {
+                override fun applyFix(dataContext: DataContext) {
+                    val args = opt.buildArguments.split(" ")
+                    args.forEachIndexed { index, arg ->
+                        if (arg == "-o" || arg == "-output") {
+                            options.outputFileName = args[index + 1]
+                        }
+                    }
+                    options.buildArguments = opt.buildArguments.replace("(-o|-output)\\s+\\S+\\s?".toRegex(), "")
+                }
+            })
+        }
     }
 
     override fun getState(executor: Executor, executionEnvironment: ExecutionEnvironment): RunProfileState? {
-        return VlangRunConfigurationRunState(executionEnvironment, this)
+        return VlangRunConfigurationRunState(executionEnvironment, this.expandedOptions())
     }
 
-    private fun Element.writeString(name: String, value: String) {
-        val opt = Element("option")
-        opt.setAttribute("name", name)
-        opt.setAttribute("value", value)
-        addContent(opt)
+    fun expandedOptions(): VlangRunConfigurationOptions.ExpandedOptions {
+        return options.expandOptions(project)
     }
-
-    private fun Element.readString(name: String): String? =
-        children
-            .find { it.name == "option" && it.getAttributeValue("name") == name }
-            ?.getAttributeValue("value")
-
-    private fun Element.writeBool(name: String, value: Boolean) {
-        writeString(name, value.toString())
-    }
-
-    private fun Element.readBool(name: String): Boolean? =
-        readString(name)?.toBoolean()
 }