+++ src/main/kotlin/io/vlang/ide/run/VlangRunConfigurationType.kt
@@ -6,22 +6,29 @@ import com.intellij.openapi.project.Project
 import io.vlang.ide.ui.VIcons
 
 class VlangRunConfigurationType : ConfigurationTypeBase(
-    ID, "V Build",
+    ID,
+    "V",
     "Run V project",
     VIcons.V
 ) {
+
     companion object {
-        const val ID = "VlangRun"
+        const val ID = "VlangRunType"
+        const val FACTORY_RUN_ID = "VlangRun"
     }
 
     init {
         addFactory(object : ConfigurationFactory(this) {
-            override fun getId() = ID
+            override fun getId() = FACTORY_RUN_ID
 
             override fun createTemplateConfiguration(project: Project) =
-                VlangRunConfiguration(project, this, "Run V project")
+                VlangRunConfiguration(project, this, null)
 
             override fun getOptionsClass() = VlangRunConfigurationOptions::class.java
+
+            override fun getName(): String = "V Build"
+
+            override fun isEditableInDumbMode(): Boolean = true
         })
     }
 }