+++ CHANGELOG.md
@@ -8,6 +8,12 @@
 - Fix IndexOutOfBoundsException while typing `}` at the end of the document
 - Change order of default value and attribute. Fixes [#22](https://github.com/vlang/intellij-v/issues/22)
 - Builtin functions will not show private members in code completion. Fixes [#9](https://github.com/vlang/intellij-v/issues/9)
+- Refactor RunConfiguration:
+  - Add validation
+  - Use relative paths to working directory
+  - Use `directory` as default run kind
+  - Expand environment variables and pass to build and run/debug configuration
+  - Add macro support
 
 ## [2024.3.1] - 2024-11-14